+++ packages/runtime/plugin-runtime/src/router/cli/index.ts
@@ -1,16 +1,15 @@
-import path from 'path';
 import {
   getEntryOptions,
   createRuntimeExportsUtils,
   isRouterV5 as isV5,
 } from '@modern-js/utils';
-import { cloneDeep } from '@modern-js/utils/lodash';
 import { ServerRoute } from '@modern-js/types';
 import type { CliPlugin, AppTools } from '@modern-js/app-tools';
-import { isRouteEntry, modifyEntrypoints } from './entry';
-import { isPageComponentFile } from './code/utils';
+import { isRouteEntry } from './entry';
+import { handleFileChange, handleModifyEntrypoints } from './handler';
 
-export { isRouteEntry, modifyEntrypoints } from './entry';
+export { isRouteEntry } from './entry';
+export { handleFileChange, handleModifyEntrypoints } from './handler';
 
 const PLUGIN_IDENTIFIER = 'router';
 
@@ -23,7 +22,6 @@ export const routerPlugin = (): CliPlugin<AppTools<'shared'>> => ({
     const runtimeConfigMap = new Map<string, any>();
 
     let pluginsExportsUtils: any;
-    let originEntrypoints: any[] = [];
 
     return {
       checkEntryPoint({ path, entry }) {
@@ -54,13 +52,7 @@ export const routerPlugin = (): CliPlugin<AppTools<'shared'>> => ({
         };
       },
       async modifyEntrypoints({ entrypoints }) {
-        const newEntryPoints = modifyEntrypoints(entrypoints);
-        const appContext = api.useAppContext();
-        const resolvedConfig = api.useResolvedConfigContext();
-        appContext.entrypoints = newEntryPoints;
-        originEntrypoints = cloneDeep(newEntryPoints);
-        const { generateCode } = await import('./code');
-        await generateCode(appContext, resolvedConfig, entrypoints, api);
+        const newEntryPoints = await handleModifyEntrypoints(api, entrypoints);
         return { entrypoints: newEntryPoints };
       },
       modifyEntryImports({ entrypoint, imports }: any) {
@@ -135,33 +127,7 @@ export const routerPlugin = (): CliPlugin<AppTools<'shared'>> => ({
         }
       },
       async fileChange(e) {
-        const appContext = api.useAppContext();
-        const { appDirectory, entrypoints } = appContext;
-        const { filename, eventType } = e;
-        const nestedRouteEntries = entrypoints
-          .map(point => point.nestedRoutesEntry)
-          .filter(Boolean) as string[];
-        const pagesDir = entrypoints
-          .map(point => point.entry)
-          // should only watch file-based routes
-          .filter(entry => entry && !path.extname(entry))
-          .concat(nestedRouteEntries);
-        const isPageFile = (name: string) =>
-          pagesDir.some(pageDir => name.includes(pageDir));
-
-        const absoluteFilePath = path.resolve(appDirectory, filename);
-        const isRouteComponent =
-          isPageFile(absoluteFilePath) && isPageComponentFile(absoluteFilePath);
-
-        if (
-          isRouteComponent &&
-          (eventType === 'add' || eventType === 'unlink')
-        ) {
-          const resolvedConfig = api.useResolvedConfigContext();
-          const { generateCode } = await import('./code');
-          const entrypoints = cloneDeep(originEntrypoints);
-          await generateCode(appContext, resolvedConfig, entrypoints, api);
-        }
+        await handleFileChange(api, e);
       },
     };
   },
